+++ src/org/xjge/components/Component.java
@@ -1,0 +1,12 @@
+package org.xjge.components;
+
+/**
+ * 
+ * @author J Hoffman
+ * @since 4.0.0
+ */
+public sealed interface Component permits AxisComponent, ButtonComponent, CursorComponent, ScrollWheelComponent {
+    
+    abstract float getInputValue();
+    
+}