+++ src/org/xjge/components/ScrollWheelComponent.java
@@ -1,0 +1,15 @@
+package org.xjge.components;
+
+/**
+ * 
+ * @author J Hoffman
+ * @since 4.0.0
+ */
+public final class ScrollWheelComponent implements Component {
+
+    @Override
+    public float getInputValue() {
+        return 0;
+    }
+
+}